Revert test changes
